# BarToYou-WebApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BarToYou API",
        "description": "Documentación Swagger de la API de BarToYou",
        "contact": {
            "email": "soporte@bartoyou.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/bartoyou/consumption-categories": {
            "get": {
                "tags": [
                    "Consumption Categories"
                ],
                "summary": "Listar categorías de consumo",
                "operationId": "43c5db2631fa324a9f02db1bbfb734da",
                "responses": {
                    "200": {
                        "description": "Listado paginado de categorías"
                    }
                }
            },
            "post": {
                "tags": [
                    "Consumption Categories"
                ],
                "summary": "Crear una nueva categoría de consumo",
                "operationId": "410ff12bd133f88a88b7429d838114eb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "nombre"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Bebidas"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoría creada con éxito"
                    },
                    "400": {
                        "description": "La categoría ya existe"
                    }
                }
            }
        },
        "/api/bartoyou/consumption-categories/{id}": {
            "get": {
                "tags": [
                    "Consumption Categories"
                ],
                "summary": "Obtener una categoría de consumo por ID",
                "operationId": "7d33ee02665555f5eba5ab464644602c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría encontrada"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Consumption Categories"
                ],
                "summary": "Actualizar una categoría de consumo",
                "operationId": "fa21baea8df73ca06e22a68f91d480a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Comida"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría actualizada correctamente"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Consumption Categories"
                ],
                "summary": "Eliminar una categoría de consumo",
                "operationId": "82e8ac1d350422090525e0cf970fe674",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminación completada"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            }
        },
        "/api/bartoyou/consumptions": {
            "get": {
                "tags": [
                    "Consumptions"
                ],
                "summary": "Listar todos los consumos",
                "operationId": "ad81c93ce48703764dee49d6d72f6d88",
                "responses": {
                    "200": {
                        "description": "Listado de consumos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Consumptions"
                ],
                "summary": "Crear un nuevo consumo",
                "operationId": "db6fb90d5ea83444f6f5974f1671689c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category_id",
                                    "image"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Coca-Cola"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consumo creado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/bartoyou/consumptions/{id}": {
            "get": {
                "tags": [
                    "Consumptions"
                ],
                "summary": "Obtener un consumo por ID",
                "operationId": "dda30765f4b713e28b9c510bf036b7b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del consumo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consumo encontrado"
                    },
                    "404": {
                        "description": "Consumo no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Consumptions"
                ],
                "summary": "Actualizar un consumo",
                "operationId": "1eba700c0f08455f80697ebaf9d78c19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del consumo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pepsi"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consumo actualizado correctamente"
                    },
                    "404": {
                        "description": "Consumo no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Consumptions"
                ],
                "summary": "Eliminar un consumo",
                "operationId": "c6f1b5ee2f1fa79364e0d834b171496d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del consumo a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consumo eliminado correctamente"
                    },
                    "404": {
                        "description": "Consumo no encontrado"
                    }
                }
            }
        },
...
    },
...
}